Refactor: makeAssistant 관련 코드를 features 내부로 이동 (#46)
