+++ src/app/(app)/make-assistant.tsx
@@ -3,7 +3,7 @@ import { useLocalSearchParams } from "expo-router";
 import { getMakeAssistantQuestionFn } from "@/src/apis/aiProfileApi";
 import Loading from "@/src/components/ui/Loading";
 import CommonError from "@/src/components/ui/CommonError";
-import MakeAssistantPage from "@/src/pages/MakeAssistant";
+import MakeAssistantContainer from "@/src/features/makeAssistant/containers/MakeAssistantContainer";
 
 /**
  * @description 어시스턴트 생성하는 페이지의 라우터
@@ -15,8 +15,7 @@ function MakeAssistantRouter() {
 
   // 어시스턴트 복제인 경우만 실행되는 query, 복제인 경우 원본의 answer를 받아옴
   const { isPending, isError, data, refetch } = useQuery({
-    queryFn: () =>
-      aiProfileId ? getMakeAssistantQuestionFn(aiProfileId) : Promise.reject("No aiProfileId"),
+    queryFn: () => (aiProfileId ? getMakeAssistantQuestionFn(aiProfileId) : Promise.reject("No aiProfileId")),
     queryKey: ["make-assistant-question", aiProfileId ?? ""],
     enabled: !!aiProfileId,
   });
@@ -30,9 +29,9 @@ function MakeAssistantRouter() {
       return <CommonError titleText="에러 발생" buttonText="재시도" onPress={() => refetch()} />;
     }
 
-    return <MakeAssistantPage prevAnswer={data.result} />;
+    return <MakeAssistantContainer prevAnswer={data.result} />;
   }
 
-  return <MakeAssistantPage />;
+  return <MakeAssistantContainer />;
 }
 export default MakeAssistantRouter;
